move PendingEnvelopes from herder to overlay
