+++ src/herder/HerderImpl.cpp
@@ -8,7 +8,6 @@
 #include "crypto/SHA.h"
 #include "herder/HerderPersistence.h"
 #include "herder/LedgerCloseData.h"
-#include "herder/PendingEnvelopes.h"
 #include "herder/TxSetFrame.h"
 #include "item/ItemKey.h"
 #include "ledger/LedgerManager.h"
@@ -17,6 +16,7 @@
 #include "main/Config.h"
 #include "main/PersistentState.h"
 #include "overlay/OverlayManager.h"
+#include "overlay/PendingEnvelopes.h"
 #include "overlay/QSetCache.h"
 #include "overlay/TxSetCache.h"
 #include "scp/LocalNode.h"
