+++ src/main/ApplicationImpl.cpp
@@ -16,7 +16,6 @@
 #include "database/Database.h"
 #include "herder/HerderImpl.h"
 #include "herder/HerderPersistence.h"
-#include "herder/PendingEnvelopes.h"
 #include "history/HistoryArchiveManager.h"
 #include "history/HistoryManager.h"
 #include "invariant/AccountSubEntriesCountIsValid.h"
@@ -39,6 +38,7 @@
 #include "medida/reporting/console_reporter.h"
 #include "medida/timer.h"
 #include "overlay/OverlayManager.h"
+#include "overlay/PendingEnvelopes.h"
 #include "process/ProcessManager.h"
 #include "scp/LocalNode.h"
 #include "scp/QuorumSetUtils.h"
