+++ src/overlay/EnvelopeItemMap.cpp
@@ -2,7 +2,7 @@
 // under the Apache License, Version 2.0. See the COPYING file at the root
 // of this distribution or at http://www.apache.org/licenses/LICENSE-2.0
 
-#include "herder/EnvelopeItemMap.h"
+#include "overlay/EnvelopeItemMap.h"
 
 namespace stellar
 {
