+++ src/overlay/FetchingEnvelopes.h
@@ -4,8 +4,8 @@
 // under the Apache License, Version 2.0. See the COPYING file at the root
 // of this distribution or at http://www.apache.org/licenses/LICENSE-2.0
 
-#include "herder/EnvelopeItemMap.h"
-#include "herder/NodesInQuorum.h"
+#include "overlay/EnvelopeItemMap.h"
+#include "overlay/NodesInQuorum.h"
 #include "xdr/Stellar-SCP.h"
 
 #include <map>
