+++ src/overlay/NodesInQuorum.cpp
@@ -2,8 +2,7 @@
 // under the Apache License, Version 2.0. See the COPYING file at the root
 // of this distribution or at http://www.apache.org/licenses/LICENSE-2.0
 
-#include "herder/NodesInQuorum.h"
-#include "herder/Herder.h"
+#include "overlay/NodesInQuorum.h"
 #include "main/Application.h"
 #include "scp/SCP.h"
