+++ src/overlay/PendingEnvelopes.cpp
@@ -2,7 +2,7 @@
 // under the Apache License, Version 2.0. See the COPYING file at the root
 // of this distribution or at http://www.apache.org/licenses/LICENSE-2.0
 
-#include "herder/PendingEnvelopes.h"
+#include "overlay/PendingEnvelopes.h"
 #include "main/Application.h"
 #include "overlay/OverlayManager.h"
 #include "overlay/QSetCache.h"
