+++ src/overlay/PendingEnvelopes.h
@@ -4,9 +4,9 @@
 // under the Apache License, Version 2.0. See the COPYING file at the root
 // of this distribution or at http://www.apache.org/licenses/LICENSE-2.0
 
-#include "herder/FetchingEnvelopes.h"
 #include "herder/Herder.h"
-#include "herder/ReadyEnvelopes.h"
+#include "overlay/FetchingEnvelopes.h"
+#include "overlay/ReadyEnvelopes.h"
 #include "xdr/Stellar-types.h"
 
 /**
