+++ src/overlay/ReadyEnvelopes.cpp
@@ -2,7 +2,7 @@
 // under the Apache License, Version 2.0. See the COPYING file at the root
 // of this distribution or at http://www.apache.org/licenses/LICENSE-2.0
 
-#include "herder/ReadyEnvelopes.h"
+#include "overlay/ReadyEnvelopes.h"
 #include "herder/Herder.h"
 #include "main/Application.h"
 #include "scp/SCP.h"
