+++ src/overlay/ReadyEnvelopesTests.cpp
@@ -2,8 +2,8 @@
 // under the Apache License, Version 2.0. See the COPYING file at the root
 // of this distribution or at http://www.apache.org/licenses/LICENSE-2.0
 
-#include "herder/ReadyEnvelopes.h"
 #include "main/Application.h"
+#include "overlay/ReadyEnvelopes.h"
 #include "test/TestPrinter.h"
 #include "test/TestUtils.h"
 #include "test/test.h"
